Saved distribution of power injection node degree
The file power_injection_degree.mat contains the poisson
random variable that describe the degree where injection nodes
occure. In addition the a variable with the percent of feeders
where the root (node at h=1) has an injection is saved. This
is expected to be used to insure that the root is not always
given a power injection