Cleaned up code and split into functions
The various testing/ploting functions were split into their own
functions with the ability to output a figure or not.
A testing function was created that aggregates all of these
data import scrip was created that imports all the distributions etc.
Excecution now takes place via a main script that calls the
single_feeder_gen function with ''user'' inputs and returns
the node and edge tables. The testing function is then used
to evaluate the feeders