Updates to input and new data saved
The input_2.csv file was recreated to insure that all feeders
have more than 1 node. Otherwise the algorithm appears to be
stuck in an infinit loop.

pmax_coeff.mat is added which contains the coefficients for a
power law curve to limit the amount of load assigned to a node
at a given hop distance.