Added cap to load capacity
Load capacity is now enforced based on hop distance.
Also added a function that plots the maximum load at each
hop distance vs. the cap function.
Note: since there is a renormalizaton step the load CAN exceed
the threshold set for it. In fact it does.
Finally, addressed a small error in how the downstream power
plot was being created to sum the total load only over positive P
this way the maximum is as expected, 1.