Improve load cap handling
Following the normalization the load cap is again enforced
so that no exceesive loads are generated.
The error between the inputs and the eventual output of
Stotal, Pinj_total as well as the calculated Ptotal is provided
as a new output of the single_feeder_gen function.
Initial testing shows that the error is quite small