Reverse Flow Mitigation Implementation
Optional input to allow for reverse flow mitigation is passed.
The default is on. This procedure checks if there are negative
power flows and then tries to swap the injection nodes with
other load nodes to alleviate the situation.
To do this the downstream power calculation needs to be performed
regularly, so it was wraped into a function.
Swapping is also wrapped into a function since we occasionally
need to swap nodes back if the results are undesierable.
This mitigation will not always work.But it works often.