Distribution Transformer Library
Created a distribution transformer library similar to the cable type
library based on a k-means cluster on the distribution transformers in
the data. The library was created from 6 clusters roughly centered
around 15, 20, 30, 40, 60, and 80 MVA. The 2 samples closest to the
cluster center are taken, as well as the furthest 2 samples one with
rating below the centroid and one with rating above.
 Changes to be committed:
	new file:   data/DistributionTransformerLibrary.csv