Modification of transformer selection process
The estimated voltage drop is now calculated using the transformer
impedance and estimated power in the transformer. This term is added to
the minimization, which is now carried out in per-unit.

IMPORTANT: the conversion to per-unit is done using the 10 MVA base.
This is a somewhat arbitrary choice but is consistent with the data
coming from Vision and the way the matpower case is created. This WILL
have a difference if changed since the estimated S will decrease and
	deltaV = sqrt(Sest*Zxfrm) will decrease.
 Changes to be committed:
	modified:   single_feeder_gen.m