Alter matpower output to match new xfrm modelling
The impedance of the transformer is now correctly calculated accounting
for the system data. Additionally, a tap is specified based on the
transformer hv rating and the system voltage.
This should be considered as a changable parameter, but the setting
should be more appropriate initially.
Based on the assumptions made the voltage of the hv bus (also the slack
bus) is set to 1 pu, rather than the higher set point it previously had.
 Changes to be committed:
	modified:   matpower_fmt.m