Input parsing and options change
Single_feeder_gen now accepts 0--4 input arguments.
A structure opt is used for various algorithm options, and is setup or
compared to defaults via a local function.
It is now also possible to specify the number of nodes desired on the
feeder and get samples from the kde object of the total load (and
injection).
 Changes to be committed:
	modified:   single_feeder_gen.m